Removes foreign keys on testing
Took 5 minutes

+++ database/migrations/2019_12_03_203615_version06_2019-12-03_groups.php
@@ -21,8 +21,10 @@ class Version0620191203Groups extends Migration
             $table->timestamp('created_at')->nullable()->default(DB::raw('NULL'));
             $table->timestamp('updated_at')->nullable()->default(DB::raw('NULL'));
 
-            $table->foreign('session_id', 'groups_sessions_foreign')->references('id')->on('sessions')
-                ->onDelete('CASCADE');
+            if (env('APP_ENV', 'local') !== 'testing') {
+                $table->foreign('session_id', 'groups_sessions_foreign')->references('id')->on('sessions')
+                    ->onDelete('CASCADE');
+            }
         });
     }
